+++ CHANGELOG.md
@@ -25,6 +25,8 @@
   - `cider-completion-annotations-alist` controls the abbreviations used in annotations.
   - `cider-completion-annotations-include-ns` controls when to include the candidate namespace in annotations.
 * Inspector middleware now relies on `eval` middleware, adding support for ClojureScript.
+* Better printing of large amounts of exception cause data in the error buffer.
+  - New defcustom, `cider-stacktrace-print-length` (boolean).
 
 ### Changes
 
@@ -46,6 +48,7 @@ when in buffer that's not visiting a file (e.g. a REPL buffer).
 * [#977](https://github.com/clojure-emacs/cider/issues/977): `cider-format-region` now respects indentation of the region start position.
 * [#979](https://github.com/clojure-emacs/cider/issues/979): Fixed the inspector buffer popping up needlessly.
 * [#981](https://github.com/clojure-emacs/cider/issues/981): Updated `cider-find-file` to use `find-buffer-visiting` instead of `get-file-buffer`.
+* [#1004](https://github.com/clojure-emacs/cider/issues/1004): `:repl-env` key is now filtered from exception causes, as it contains unprintably large strings of compiled javascript when using ClojureScript.
 
 ## 0.8.2 / 2014-12-21